1. Revert UpsertStreamTableSink, RetractStreamTableSink 2. Move valuesInputFormat to values subpackage 3. Minor update in StreamExecSink and BatchExecSink
